+++ lib/glimmer/libui/control_proxy/date_time_picker_proxy.rb
@@ -69,4 +69,4 @@ module Glimmer
   end
 end
 
-Dir[File.expand_path('./date_time_picker_proxy/*.rb', __dir__)].each {|f| require f}
+Dir[File.expand_path("./#{File.basename(__FILE__, '.rb')}/*.rb", __dir__)].each {|f| require f}