+++ lib/glimmer/libui/control_proxy/menu_item_proxy.rb
@@ -42,4 +42,4 @@ module Glimmer
   end
 end
 
-Dir[File.expand_path('./menu_item_proxy/*.rb', __dir__)].each {|f| require f}
+Dir[File.expand_path("./#{File.basename(__FILE__, '.rb')}/*.rb", __dir__)].each {|f| require f}