+++ lib/glimmer/libui/control_proxy/message_box.rb
@@ -22,30 +22,13 @@
 module Glimmer
   module LibUI
     class ControlProxy
-      # Common logic for all column proxy objects
-      module Column
-        def initialize(keyword, parent, args, &block)
-          @keyword = keyword
-          @parent_proxy = parent
-          @args = args
-          @block = block
-          @enabled = true
-          post_add_content if @block.nil?
-        end
-        
-        def name
-          @args.first
-        end
-        
-        # column index used in table append column API call
-        # expanded to ensure DualColumn index accounts for two columns acting as one
-        def column_index
-          @column_index ||= @parent_proxy.send(:next_column_index)
-        end
-        
-        # actual index within table columns (disregarding nil fillings that account for DualColumn instances)
-        def index
-          @parent_proxy.columns.compact.index(self)
+      # Represents LibUI message box objects
+      module MessageBox
+        def build_control
+          @args.prepend ControlProxy.main_window_proxy if @args[0].nil? || @args[0].is_a?(String)
+          @args[1] ||= ''
+          @args[2] ||= ''
+          super
         end
       end
     end