+++ lib/glimmer/libui/control_proxy/multiline_entry_proxy.rb
@@ -36,4 +36,4 @@ module Glimmer
   end
 end
 
-Dir[File.expand_path('./multiline_entry_proxy/*.rb', __dir__)].each {|f| require f}
+Dir[File.expand_path("./#{File.basename(__FILE__, '.rb')}/*.rb", __dir__)].each {|f| require f}